#228 Delete marked files flow

+++ backend/src/infraestructure/adapters/repositories/MysqlFileRepository.ts
@@ -26,8 +26,7 @@ export class MysqlFileRepository implements FileRepository {
     try {
       const [rows] = await promisePool.execute(query, [sectionId]);
       return (rows as any[]).map(
-        (row) =>
-          new UserFile(row.public_id, row.url, row.section_id)
+        (row) => new UserFile(row.public_id, row.url, row.section_id)
       );
     } catch (error) {
       console.error("Error retrieving files by section:", error);
@@ -35,16 +34,19 @@ export class MysqlFileRepository implements FileRepository {
     }
   }
 
-  async delete(publicId: string): Promise<void> {
+  async deleteFilesByPublicIds(publicIds: string[]): Promise<void> {
+    if (publicIds.length === 0) return;
+
+    const placeholders = publicIds.map(() => "?").join(", ");
     const query = `
-      DELETE FROM users_files
-      WHERE public_id = ?`;
+    DELETE FROM users_files
+    WHERE public_id IN (${placeholders})`;
 
     try {
-      await promisePool.execute(query, [publicId]);
+      await promisePool.execute(query, publicIds);
     } catch (error) {
-      console.error("Error deleting file:", error);
-      throw new Error("Could not delete file");
+      console.error("Error deleting multiple files:", error);
+      throw new Error("Could not delete files");
     }
   }
 }
